+++ src/shared/StyledComponents.tsx
@@ -1,0 +1,11 @@
+import styled from "styled-components";
+
+export const InputFieldStyled = styled.input`
+  border-radius: 4px;
+  border: 1px solid hsl(0, 0%, 80%);
+  width: 300px;
+  height: 38px;
+  font-size: 14px;
+  padding-left: 10px;
+  box-sizing: border-box;
+`;
